requirements.txt
flask==2.3.3
gunicorn==21.2.0
openai==1.3.0
python-dotenv==1.0.0
requests==2.31.0
APScheduler==3.10.4
supabase==1.0.3
python-twitter-v2==0.8.1 